+++ client/launcher.py
@@ -242,6 +242,10 @@ parser_list_tasks.add_argument('-q', '--quiet', action='store_true',
                                help='only display task_id')
 parser_list_tasks.add_argument('-S', '--status',
                                help='filter on status value')
+parser_list_tasks.add_argument('-se', '--service',
+                               help='filter on service')
+parser_list_tasks.add_argument('-pa', '--parent', action='store_true',
+                               help='display parent task')
 
 parser_del_tasks = subparsers_tasks.add_parser('delete',
                                                help='{dt} delete tasks matching prefix pattern')
@@ -252,6 +256,8 @@ parser_del_tasks.add_argument('-p', '--prefix', required=True,
 parser_status = subparsers_tasks.add_parser('status',
                                             help='{status} get status of a task')
 shortcut_map["status"] = ["task", "status"]
+shortcut_map["parent"] = ["task", "parent"]
+shortcut_map["service"] = ["task", "service"]
 parser_status.add_argument('task_id', help='task identifier')
 
 parser_terminate = subparsers_tasks.add_parser('terminate',
@@ -358,19 +364,22 @@ def process_request(serviceList, cmd, subcmd, is_json, args, auth=None):
             if len(busymsg):
                 res += "\n" + "\n".join(busymsg)
     elif cmd == "task" and subcmd == "list":
-        r = requests.get(os.path.join(args.url, "task/list", args.prefix + '*'), auth=auth)
+        r = requests.get(os.path.join(args.url, "task/list", args.prefix + '*'), auth=auth,
+                         params={"with_parent": args.parent, "service": args.service, "status": args.status})
         if r.status_code != 200:
             raise RuntimeError('incorrect result from \'task/list\' service: %s' % r.text)
         result = r.json()
-        if args.status:
-            result = [r for r in result if r["status"] == args.status]
+
         if args.quiet:
             res = []
             for k in result:
                 res.append(k["task_id"])
         elif not is_json:
-            res = PrettyTable(["Task ID", "Resource", "Priority",
-                               "Launch Date", "Image", "Status", "Message"])
+            headers= ["Task ID", "Service", "Resource", "Priority", "Launch Date", "Image", "Status", "Message" ]
+            if args.parent:
+                headers.append("parent")
+
+            res = PrettyTable(headers)
             res.align["Task ID"] = "l"
             for k in sorted(result, key=lambda k: float(k.get("launched_time") or 0)):
                 date = datetime.fromtimestamp(
@@ -383,8 +392,12 @@ def process_request(serviceList, cmd, subcmd, is_json, args, auth=None):
                 if p != -1:
                     k["image"] = k["image"][p+1:]
                 task_id = k["task_id"]
-                res.add_row([task_id, resource, int(k["priority"] or 0),
-                             date, k["image"], k["status"], k.get("message")])
+                if args.parent:
+                    res.add_row([task_id, k["service"], resource, int(k["priority"] or 0),
+                                 date, k["image"], k["status"], k.get("message"), k.get("parent")])
+                else:
+                    res.add_row([task_id, k["service"], resource, int(k["priority"] or 0),
+                                 date, k["image"], k["status"], k.get("message")])
         else:
             res = r.json()
     elif cmd == "service" and subcmd == "describe":